feat: change settings activity to fragments
