+++ app/build.gradle.kts
@@ -8,20 +8,20 @@ plugins {
 android {
     namespace = "com.sparkleside"
     compileSdk = 34
-    
+
     defaultConfig {
         applicationId = "com.sparkleside"
         minSdk = 24
         targetSdk = 34
         versionCode = 1
         versionName = "1.0"
-      
-        vectorDrawables { 
+
+        vectorDrawables {
             useSupportLibrary = true
         }
         resConfigs("ar-rTN", "es-rES", "fr-rTN", "in-rID", "pt-rBR", "pt-rPT", "zh-rCN")
     }
-    
+
     compileOptions {
         sourceCompatibility = JavaVersion.VERSION_17
         targetCompatibility = JavaVersion.VERSION_17
@@ -30,9 +30,9 @@ android {
 
     buildFeatures {
         viewBinding = true
-        buildConfig= true
+        buildConfig = true
     }
-    
+
     signingConfigs {
         create("release") {
             // temporary keystore
@@ -49,9 +49,9 @@ android {
             keyPassword = "testkey"
         }
     }
-    
+
     buildTypes {
-   getByName("release"){
+        getByName("release") {
             isMinifyEnabled = false
             isShrinkResources = false
             proguardFiles(getDefaultProguardFile("proguard-android-optimize.txt"), "proguard-rules.pro")
@@ -67,13 +67,13 @@ tasks.withType<org.jetbrains.kotlin.gradle.tasks.KotlinCompile> {
 }
 
 dependencies {
-    implementation("androidx.constraintlayout:constraintlayout:2.1.4")
-    implementation("androidx.appcompat:appcompat:1.6.1")
+    implementation("androidx.constraintlayout:constraintlayout:2.2.0")
+    implementation("androidx.appcompat:appcompat:1.7.0")
     implementation("com.google.android.material:material:1.13.0-alpha08")
     implementation(platform("io.github.Rosemoe.sora-editor:bom:0.23.4"))
     implementation("io.github.Rosemoe.sora-editor:editor:0.23.4")
     implementation("com.github.bumptech.glide:glide:4.16.0")
-    implementation("androidx.activity:activity:1.6.0-alpha05")
+    implementation("androidx.activity:activity:1.9.3")
     implementation(fileTree(mapOf("dir" to "libs", "include" to listOf("*.jar", "*.aar"))))
     implementation("com.blankj:utilcodex:1.31.1")
     implementation("nl.dionsegijn:konfetti-xml:2.0.4")
@@ -81,9 +81,11 @@ dependencies {
     implementation(project(":peekandpop"))
     implementation(project(":maskable"))
     implementation(project(":ui-utils"))
-    
+    implementation("androidx.navigation:navigation-fragment-ktx:2.8.4")
+    implementation("androidx.navigation:navigation-ui-ktx:2.8.4")
+
     coreLibraryDesugaring("com.android.tools:desugar_jdk_libs:2.1.2")
-    
+
     //termux
     implementation("com.github.termux.termux-app:terminal-view:0.118.1")
     implementation("com.github.termux.termux-app:terminal-emulator:0.118.1")
