+++ app/src/main/AndroidManifest.xml
@@ -1,6 +1,7 @@
 <?xml version="1.0" encoding="utf-8"?>
 
-<manifest xmlns:android="http://schemas.android.com/apk/res/android">
+<manifest xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:tools="http://schemas.android.com/tools">
 
     <uses-permission android:name="android.permission.INTERNET" />
 
@@ -13,7 +14,8 @@
         android:localeConfig="@xml/local_config"
         android:roundIcon="@mipmap/ic_launcher"
         android:supportsRtl="false"
-        android:theme="@style/Theme.Sparkles">
+        android:theme="@style/Theme.Sparkles"
+        tools:targetApi="tiramisu">
 
         <activity
             android:name="com.sparkleside.ui.activities.MainActivity"
@@ -25,21 +27,17 @@
         </activity>
         <activity
             android:name=".ui.activities.TerminalActivity"
-            android:exported="true"
-            android:windowSoftInputMode="adjustResize" 
-            android:enableOnBackInvokedCallback="true"/>
+            android:exported="false"
+            android:windowSoftInputMode="adjustResize" />
 
         <activity
             android:name="com.sparkleside.ui.activities.SettingsActivity"
-            android:exported="true"
-            android:enableOnBackInvokedCallback="true" />
+            android:exported="false" />
         <activity
             android:name="com.sparkleside.ui.activities.AboutActivity"
-            android:exported="true" 
-            android:enableOnBackInvokedCallback="true"/>
+            android:exported="false" />
         <activity
             android:name="com.sparkleside.ui.activities.AppearanceActivity"
-            android:exported="true"
-            android:enableOnBackInvokedCallback="true" />
+            android:exported="false" />
     </application>
 </manifest>
