+++ app/src/main/java/com/sparkleside/preferences/Preferences.java
@@ -1,7 +1,7 @@
 package com.sparkleside.preferences;
 
 import android.content.Context;
-import android.content.SharedPreferences;
+
 import androidx.appcompat.app.AppCompatDelegate;
 
 import com.sparkleside.BuildConfig;
