+++ app/src/main/java/com/sparkleside/ui/base/BaseActivity.kt
@@ -5,10 +5,9 @@ import android.os.Bundle
 import androidx.activity.enableEdgeToEdge
 import androidx.appcompat.app.AppCompatActivity
 
-open class BaseActivity : AppCompatActivity(){
-
-        override fun onCreate(saved: Bundle?) {
-                super.onCreate(saved)
-                enableEdgeToEdge()
-        }
+open class BaseActivity : AppCompatActivity() {
+    override fun onCreate(saved: Bundle?) {
+        super.onCreate(saved)
+        enableEdgeToEdge()
+    }
 }
