+++ app/src/main/res/anim/exit_transition.xml
@@ -1,0 +1,11 @@
+<?xml version="1.0" encoding="utf-8"?>
+<set xmlns:android="http://schemas.android.com/apk/res/android">
+    <alpha
+        android:duration="200"
+        android:fromAlpha="1.0"
+        android:toAlpha="0.0" />
+    <translate
+        android:duration="200"
+        android:fromXDelta="0%"
+        android:toXDelta="-50%" />
+</set>
