+++ app/src/main/res/anim/pop_enter_transition.xml
@@ -1,0 +1,11 @@
+<?xml version="1.0" encoding="utf-8"?>
+<set xmlns:android="http://schemas.android.com/apk/res/android">
+    <alpha
+        android:duration="250"
+        android:fromAlpha="0.0"
+        android:toAlpha="1.0" />
+    <translate
+        android:duration="250"
+        android:fromXDelta="-50%"
+        android:toXDelta="0%" />
+</set>
